% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_peak.R
\name{epidemic_peak}
\alias{epidemic_peak}
\title{Summarize time and size of epidemic peak}
\usage{
epidemic_peak(simulation_data)
}
\arguments{
\item{simulation_data}{Simulated epidemics from metapopulation SEIR model}
}
\value{
summary
}
\description{
Summarize time and size of epidemic peak
}
