% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peak_summary.R
\name{plot_peak_summary}
\alias{plot_peak_summary}
\title{Title}
\usage{
plot_peak_summary(simulation_data)
}
\arguments{
\item{simulation_data}{}
}
\value{
Plot summarising times and size of peaks in each patch
}
\description{
Title
}
